- Moved unused classes to an example folder to optimize teamcode package - Added skeleton class for Autonomous and non-practice TeleOp - Changed the telemetry caption in PracticeOp
