+++ TeamCode/src/main/java/ExampleCode/Auto.java
@@ -1,44 +1,42 @@
-package org.firstinspires.ftc.teamcode;
-
+package ExampleCode;
+import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.Servo;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.util.ElapsedTime;
 import com.qualcomm.robotcore.util.Range;
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
+
+
+public class Auto extends OpMode{
 
 
 
-
-public class NewAuto extends OpMode {
 
     private static double Speed = 0.4;
     private static double InnerSpeed = Speed * ConfigValues.InsideRatio;
 
     private static double SecondsPerFoot = 0.5 / Speed;
 
-private enum State {
-    Begin,
-    Depart,
-    Turn1,
-    Straight,
-    Turn2,
-    Approach,
-    Dock
-}
+    private enum State {
+        Begin,
+        Depart,
+        Turn1,
+        Straight,
+        Turn2,
+        Approach,
+        Dock
+    }
 
     private static double DepartDist   =  0.0;
-    private static double Turn1Dist    =  ConfigValues.FeetPerCircle / 4; // 45
+    private static double Turn1Dist    =  ConfigValues.FeetPerCircle / 8; // 45
     // degrees
     private static double StraightDist =  4.7;
-    private static double Turn2Dist    =  ConfigValues.FeetPerCircle / 4; // 45
-    //degrees
+    private static double Turn2Dist    =  ConfigValues.FeetPerCircle / 8; // 45
+    // degrees
     private static double ApproachDist =  1.4;
 
     private static double DepartDur   =  DepartDist   * SecondsPerFoot;
     private static double Turn1Dur    =  Turn1Dist    * SecondsPerFoot;
     private static double StraightDur =  StraightDist * SecondsPerFoot;
-   private static double Turn2Dur    =  Turn2Dist    * SecondsPerFoot;
+    private static double Turn2Dur    =  Turn2Dist    * SecondsPerFoot;
     private static double ApproachDur =  ApproachDist * SecondsPerFoot;
 
     private static double DepartEnd   = DepartDur;
@@ -49,16 +47,11 @@ private enum State {
 
     private ElapsedTime timer = new ElapsedTime();
 
-    private DcMotor motorLeft;
-    private DcMotor motorRight;
-    //private DcMotor Flywheel;
+    private DcMotor LeftWheel;
+    private DcMotor RightWheel;
 
 
-
-   // private Servo PressThing;
-
-
-    private NewAuto.State state = NewAuto.State.Begin;
+    private State state = State.Begin;
 
     private double leftPower  = 0.0;
     private double rightPower = 0.0;
@@ -68,14 +61,14 @@ private enum State {
 
     private void setLeftPower(double pwr) {
         leftPower = Range.clip(pwr,  -1, 1);
-        motorLeft.setPower(leftPower);
-
+        LeftWheel.setPower(leftPower);
+        LeftWheel.setPower(leftPower);
     }
 
     private void setRightPower(double pwr) {
         rightPower = Range.clip(pwr,  -1, 1);
-        motorRight.setPower(rightPower);
-
+        RightWheel.setPower(rightPower);
+        RightWheel.setPower(rightPower);
     }
 
     private void setPower(double left, double right) {
@@ -84,9 +77,8 @@ private enum State {
     }
 
     private void setWheelMode(DcMotor.RunMode mode) {
-        motorRight.setMode(mode);
-        motorLeft.setMode(mode);
-
+        RightWheel.setMode(mode);
+        LeftWheel.setMode(mode);
     }
 
     private static int Clicks(double ft)
@@ -96,8 +88,8 @@ private enum State {
     private void setTarget(double leftFt, double rightFt) {
         int left  = Clicks(leftFt);
         int right = Clicks(rightFt);
-        motorLeft.setTargetPosition(left);
-        motorRight.setTargetPosition(right);
+        LeftWheel.setTargetPosition(left);
+        RightWheel.setTargetPosition(right);
 
     }
 
@@ -106,7 +98,7 @@ private enum State {
         rightPosFt += ft;
         setPower(Speed, Speed);
         setTarget(leftPosFt, rightPosFt);
-   }
+    }
 
     private void turnLeft(double ft) {
         leftPosFt  += ft * ConfigValues.InsideRatio;
@@ -122,83 +114,77 @@ private enum State {
         setTarget(leftPosFt, rightPosFt);
     }
 
-    public NewAuto() { }
+    public Auto() { }
 
-    @Override
+
     public void init() {
-        motorLeft  = hardwareMap.dcMotor.get("MotorLeft");
-        motorRight = hardwareMap.dcMotor.get("MotorRight");
-       //Flywheel = hardwareMap.dcMotor.get("Flywheel");
-       //PressThing = hardwareMap.servo.get("presser");
+        LeftWheel  = hardwareMap.dcMotor.get("MotorLeft");
+        RightWheel = hardwareMap.dcMotor.get("MotorRight");
 
 
-        motorLeft.setDirection(DcMotor.Direction.FORWARD);
-        motorRight.setDirection(DcMotor.Direction.REVERSE);
-       // Flywheel.setDirection(DcMotor.Direction.FORWARD);
-       // PressThing.setDirection(Servo.Direction.FORWARD);
+        LeftWheel.setDirection(DcMotor.Direction.FORWARD);
+        LeftWheel.setDirection(DcMotor.Direction.FORWARD);
 
 
-
-
-
+        setWheelMode(DcMotor.RunMode.RESET_ENCODERS);
     }
 
-    @Override
+
     public void start() {
         setPower(0.0, 0.0);
+        setWheelMode(DcMotor.RunMode.RUN_TO_POSITION);
         leftPosFt  = 0.0;
         rightPosFt = 0.0;
-        state = NewAuto.State.Begin;
+        state = State.Begin;
     }
 
-    @Override
+
     public void loop() {
         switch (state) {
-            //case Begin:
-               // timer.reset();
-               // move(DepartDist);
-               // state = NewAuto.State.Depart;
+            case Begin:
+                timer.reset();
+                move(DepartDist);
+                state = State.Depart;
                 // fall thru
-            //case Depart:
-                //if (timer.time() < DepartEnd)
-                   // break;
-               // turnLeft(Turn1Dist);
-               // state = NewAuto.State.Turn1;
+            case Depart:
+                if (timer.time() < DepartEnd)
+                    break;
+                turnLeft(Turn1Dist);
+                state = State.Turn1;
                 // fall thru
             case Turn1:
                 if (timer.time() < Turn1End)
                     break;
                 move(StraightDist);
-                state = NewAuto.State.Straight;
+                state = State.Straight;
                 // fall thru
             case Straight:
                 if (timer.time() < StraightEnd)
                     break;
                 turnLeft(Turn2Dist);
-                state = NewAuto.State.Turn2;
+                state = State.Turn2;
                 // fall thru
             case Turn2:
                 if (timer.time() < Turn2End)
                     break;
                 move(ApproachDist);
-                state = NewAuto.State.Approach;
+                state = State.Approach;
                 // fall thru
             case Approach:
                 if (timer.time() < ApproachEnd)
                     break;
                 setPower(0.0, 0.0);
-                state = NewAuto.State.Dock;
+                state = State.Dock;
                 // fall thru
             case Dock:
                 break;
         }
 
-        telemetry.addData("state", String.format("state: %s", state));
-        telemetry.addData("left",  String.format("%.2f", leftPosFt));
-        telemetry.addData("right", String.format("%.2f", rightPosFt));
+
     }
 
-    @Override
+
     public void stop() { setPower(0.0, 0.0); }
 
 } // AutoLongLeft
+
