+++ TeamCode/src/main/java/ExampleCode/AutoLongRight.java
@@ -1,8 +1,7 @@
-package org.firstinspires.ftc.teamcode;
+package ExampleCode;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorController;
 import com.qualcomm.robotcore.util.ElapsedTime;
 import com.qualcomm.robotcore.util.Range;
