+++ TeamCode/src/main/java/ExampleCode/ConfigValues.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package ExampleCode;
 
 /**
  * Single place for configuration parameters.
