+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/AutonomousOp.java
@@ -1,0 +1,19 @@
+package org.firstinspires.ftc.teamcode;
+
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.OpMode;
+
+@Autonomous(name = "Autonomous", group = "Pushbot")
+
+public class AutonomousOp extends OpMode {
+
+    @Override
+    public void init() {
+
+    }
+
+    @Override
+    public void loop() {
+
+    }
+}
